azure-phishing-detector/requirements.txt
requests
msal
python-dotenv
azure-ai-textanalytics
beautifulsoup4